[AN] feat(#267): repository 분리
